Tests for Zotero field presence
This test case adds a pre-structured Word XML file that contains extraneous Zotero markup. The test will succeed when these fields are removed and a plaintext reference remains in place. This may need to be changed later when reference parsing works.

As a side note, this has also incurred the implementation of a parser script that will take a pre-extracted Word folder and go from there. We could use this to separate out the stages of extraction because, at present, it's somewhat repetitious.