Remove processing of comments.
This fixes #19 by removing all processing of comments. We don't need them as they are usually just remnants of an editing process and it makes it far easier for us to detect styles if textruns.xsl can simpy select on w:rPr/w:sz, which is broken if you include comments.xml in the parsing list.