Major changes to docopt structure
The new docopt structure is much more clear, with less repetition. Instead of the myriad of options, we now use a far more simple [options] to avoid repetition.

I have also implemented an option to let us point the classifier at a pre-extracted document. This is useful for our Zotero test and also fixes #39.