Fix to linebreak handling inside elements other than p
Before, elements with breaks inside elements that are not "p" tags was causing a null reference error because the script could not find the expected enclosing tag in the stack.

This fix isolates handling to breaks within p tags. It also adds additional support and tests for the JATS break tag when inside a table cell or heading as per http://dtd.nlm.nih.gov/publishing/tag-library/3.0/index.html